requirements.txt
apturl==0.5.2
asgiref==3.6.0
asttokens==2.2.1
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
blinker==1.6.2
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.6
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2
django-crispy-forms==2.0
django-enumchoicefield==3.0.1
django-markdownx==4.0.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
exceptiongroup==1.1.1
executing==1.2.0
factory-boy==3.2.1
Faker==18.9.0
fasteners==0.14.1
filelock==3.6.0
Flask==2.3.2
future==0.18.2
gpg==1.16.0
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==23.1
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==3.12.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycups==2.0.1
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.3.1
pytest-django==4.5.2
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
systemd-python==234
texttable==1.6.4
tomli==2.0.1
traitlets==5.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.6
websocket-client==1.2.3
Werkzeug==2.3.7
xdg==5
xkit==0.0.0
zipp==1.0.0
